Updated quiz question gen and answer key
Made it so the correct answer can only be one multiple choice
option.  These will always be radio buttons, so if multiple options
could be correct, that'll be reflected by an additional option. E.g.
given options A, B, C, D if A and C are correct, there'd need to be
a fifth choice E with the prompt "A and B".

This makes it so the presentation and processing is the same for all
questions and doesn't rely on interpreting the database contents,
separating these concerns.

Also, updated test to reflect change.